Backend/requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.6.1
async-timeout==4.0.2
attrs==22.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
black==22.12.0
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
contourpy==1.0.6
cycler==0.11.0
datasets==2.8.0
dill==0.3.6
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.85.0
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
fvcore==0.1.5.post20221221
gdown==4.6.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
gunicorn==20.1.0
h11==0.14.0
huggingface-hub==0.11.1
hydra-core==1.3.1
idna==3.4
iopath==0.1.9
joblib==1.2.0
kiwisolver==1.4.4
loguru==0.6.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mongoengine==0.24.2
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==0.4.3
numpy==1.24.1
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.68
packaging==22.0
pandas==1.5.2
passlib==1.7.4
pathspec==0.10.3
Pillow==9.3.0
platformdirs==2.6.2
portalocker==2.6.0
protobuf==3.20.3
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pydantic==1.10.2
pymongo==4.2.0
pyparsing==3.0.9
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.7
PyYAML==6.0
pdf2image==1.16.3
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seqeval==1.2.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.20.4
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.1
threadpoolctl==3.1.0
timm==0.6.12
tokenizers==0.13.2
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.18.3
Werkzeug==2.2.2
xxhash==3.2.0
yacs==0.1.8
yarl==1.8.2